tidal-playlist-monitor/requirements.txt
# Core functionality
tidalapi>=0.8.9
apscheduler>=3.10.0
requests>=2.32.4
pyyaml>=6.0

# CLI & UI
typer>=0.9.0
rich>=13.0.0
coloredlogs>=15.0.1

# Notifications (cross-platform)
plyer>=2.1.0

# Windows-specific (conditional install)
pywin32>=306; platform_system=='Windows'
winotify>=1.1.0; platform_system=='Windows'

# Optional - For running as system service
python-daemon>=3.0.0; platform_system!='Windows'
